Add a generic COMMENT_MODULE (#93)
Introduce the new COMMENT_MODULE
The COMMENT_MODULE is the generic successor to the REPUTATION_MODULE. The purpose is a module to leave comments about characters, each assigned to a category. Each category has a freely configurable access level for reading and writing comments. To make the most of it, it integrates into other modules like TRADEBOT_MODULE, WHOIS_MODULE and RAID_MODULE

Features:
* Obsoletes the REPUTATION_MODULE, automatically converting old entries into comments.
* Configurable cooldown time between leaving comments about the same person
* Support for both, system categories (raid, reputation) and custom ones. Every category has an individual access level and there is no limit to the amount of categories.
* Automatically show comments for whois-requests  and tradebot postings
* Add `!raid comments` to see comments about everyone in the raid

Closes #82