Special 0x treatment
When starting a raffle, auctioning 0x an item will treat this  as
"unlimited", so everyone who joins is a winner.
This is basically the same as determining the order of something,
but people have to join themselves.
To make this more usful, the word "lootorder" will now
trigger and set the amount to 0, so lootorders have only
winners.