Support "Requires" csv keyword
We want to be able to update the CSV files dynamically.
But since they don't contain the schema anymore,
we have to have a way to depend on a certain version of
the schema for the CSV to be inserted.
For this, I introduce Requires as a keyword in CSV files to
require a certain migration to have been applied.
This way, we can probably update all CSV files
dynamically in the future.