Worldboss module (#237)
This module replaces the following external modules:

  - BIGBOSS_MODULE
  - GAUNTLET_MODULE

The new WORLDBOSS_MODULE supports the same functionality, except for subscribing to gauntlet events which I ditched, because the code was horrible, and I didn't see any reason to keep the functionality anyway.
On top of the old functionality, it supports syncing timers and gauntlet buff. The overall code is more robust and easier to maintain. Old timers are automatically converted, and so are routes.