Redo sub commands (#285)
Until now, sub-commands were defined as regular expressions. If it matched, then only the sub-commands with that regexp as "cmd" were even considered for execution. This caused several problems:
1. In order to understand which command-invocations were affected by which sub-command, you had to be able to understand regexps
2. If your sub-command had something like 'points .+', then it was impossible to to the rights of the base-command for everything but points and all other sub-commands would have to "poke holes" into the '.+' catchall like 'points log', 'points log all', even though, there was no need for a new access level for these commands.

This change removes regular expressions from sub-commands. The "cmd" of the sub-command is now a text to your choosing and all regexps of all command handlers of the base-command will now be compared against the issued command.
The only meaning of the "cmd" of a sub-command is to give all the handlers assigned to it a "group name" and their own access level.

Closes #260