Unstuck buddylist->isOnline
The buddylist should never have to lookup the
ID of its buddies, because you always get it
right with the online/offline package.
"!track online" got stuck while being run the first
time for a large batch of names, because a lot
of the people are inactive, so the first call triggered
a load of user-id-lookups which were failing anyway.