Support Discord Slash-commands (#367)
This introduces a new setting "discord_slash_commands" which allows you to have your bot register some command as slash-commands and react to them and to interaction-events like buttons.
Since Discord only allows 100 global slash-commands per application, we need to let the users choose which commands they want to see exposed as slash-commands. To manage these, I created the "!discord slash"-commands.
The bot knows 2 modes, how to reply to a slash-command:

- only to the user issuing it
- by treating the command like a regular message and routing it to the bot and answering openly

Because commands can require fixed words in certain positions ("route add foo -> bar" for example), it's not possible to make proper slash-commands out of the bot-commands, since Discord doesn't support this. Currently, we only detect, if a command has

- no arguments
- only mandatory arguments
- at least one optional argument

and set these flags accordingly when registering the slash-commands.