Fix slow "track addorg"
Inserting a "logoff" information into the tracker
table takes ~10-20ms, which adds to a lot of blocked time
when  adding big orgs. Now, we do a mass insert
of dummy-logoff when tracking a new org and skip
adding 1 dummy-entry per offline-event.
This now increased the speed by a factor of 10 to 20.